requirements.txt
streamlit==1.28.1
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.7
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-multipart==0.0.6  # For file uploads in deals
pytest==7.4.3  # For tests